Refactor OrderBook component structure and enhance UI elements
- Replaced the OrderBook component with OrderBookContainer for improved modularity.
- Introduced OrderBookTable to better organize order display functionality.
- Updated TradingContent to utilize the new OrderBookContainer, enhancing the overall layout.
- Removed deprecated OrderBook, OrderRow, and OrderBookSection components to streamline the codebase.
- Enhanced OrderBookHeader styling for better visual consistency and clarity.
