+++ pengui/src/features/trading/index.ts
@@ -9,9 +9,9 @@ export { useOrderBookFilters, OrderBookFiltersProvider } from './model/OrderBook
 export { useOrderBookOfferSubmission } from './model/useOrderBookOfferSubmission'
 
 // UI exports
-export { default as OrderBook } from './ui/orderbook/OrderBook'
+export { default as OrderBookContainer } from './ui/orderbook/OrderBookContainer'
+export { default as OrderBookTable } from './ui/orderbook/OrderBookTable'
 export { default as OrderBookFilters } from './ui/orderbook/OrderBookFilters'
-export { default as OrderRow } from './ui/orderbook/OrderRow'
 export { default as OrderTooltip } from './ui/orderbook/OrderTooltip'
 export { default as TradingLayout } from './ui/layout/TradingLayout'
 export { default as TradingContent } from './ui/layout/TradingContent'
